% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffusion.R
\name{get_infected}
\alias{get_infected}
\title{Get the infected nodes}
\usage{
get_infected(g, transmission_rate, seed_num, random_seed)
}
\arguments{
\item{g}{A network object generated using igraph.}

\item{transmission_rate}{A number.}

\item{seed_num}{The number of seeds.}

\item{random_seed}{A number for the function set.seed().}
}
\value{
a list of infected nodes (over time).
}
\description{
Get the infected nodes
}
\examples{
get_infected(barabasi.game(100), 0.4, 1, 2004)
get_infected(erdos.renyi.game(100, 0.1), 0.4, 1, 2004)
}

