% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffusion.R
\name{plot_gif}
\alias{plot_gif}
\title{plot the diffusion network diffusion (over time)}
\usage{
plot_gif(infected)
}
\arguments{
\item{infected}{A list of infected nodes (over time)}
}
\value{
several plots of diffusion networks.
}
\description{
plot the diffusion network diffusion (over time)
}
\examples{
plot_gif(infected)
plot_gif(infected[1])
}

