% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffusion.R
\name{plot_time_series}
\alias{plot_time_series}
\title{plot the time series of network diffusion}
\usage{
plot_time_series(infected)
}
\arguments{
\item{infected}{A list of infected nodes (over time)}
}
\value{
plots of two time series.
}
\description{
plot the time series of network diffusion
}
\examples{
plot_time_series(infected)
plot_time_series(infected[1:5])
}

