% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffusion.R
\name{plot_gif}
\alias{plot_gif}
\title{plot the diffusion network diffusion (over time)}
\usage{
plot_gif(infected, g)
}
\arguments{
\item{infected}{A list of infected nodes (over time)}

\item{g}{A network object generated using igraph.}
}
\value{
several plots of diffusion networks.
}
\description{
plot the diffusion network diffusion (over time)
}
\examples{
require(networkdiffusion)
require(igraph)
require(animation)

# generate a BA network
g = barabasi.game(100)
infected = get_infected(g, 0.4, 1, 2004)
plot_gif(infected[[2]], g)
plot_time_series(infected)

saveGIF({
 ani.options("convert")
 plot_gif(infected, g)
 }, interval = 0.3, movie.name = "ba.gif", ani.width = 600, ani.height = 600)
}

